+++ src/Goodness.h
@@ -6,7 +6,7 @@
 #include "Utility.h"
 using namespace std;
 
-	void calculateGoodnessGrid(Grid* topographyGrid, Grid* behaviorGrid, Grid* goodnessGrid, double bias, double range);
+	void calculateGoodnessGrid(Grid* topographyGrid, Grid* behaviorGrid, Grid* goodnessGrid, int bias, int range);
 	std::set<std::pair<int,int>> getCells(std::pair <int,int> *origin, std::pair <int,int> *target);
 	Eigen::MatrixXd calcPercentViz(Grid* topographyGrid, int rStart, int cStart, int rng);
 	void goodFish(Grid* topographyGrid, Grid* behaviorGrid, Grid* goodnessGrid, double range);
